refactor(modules): use clearly defined modules and explicit package namespace with `__all__`
