+++ src/uform/gen_model.py
@@ -3,7 +3,15 @@ from typing import List, Optional, Tuple, Union
 import torch
 import torch.nn.functional as F
 from torch import nn
-from torchvision.transforms import *
+from torchvision.transforms import (
+    CenterCrop,
+    Compose,
+    InterpolationMode,
+    Normalize,
+    RandomResizedCrop,
+    Resize,
+    ToTensor,
+)
 from transformers import AutoConfig, AutoTokenizer
 from transformers.configuration_utils import PretrainedConfig
 from transformers.modeling_outputs import CausalLMOutputWithPast
@@ -12,24 +20,19 @@ from transformers.models.auto.modeling_auto import AutoModel, AutoModelForCausal
 from transformers.processing_utils import ProcessorMixin
 from transformers.tokenization_utils_base import BatchEncoding
 
-from uform.models import VisualEncoder
+from .models.encoders import LayerScale, VisualEncoder
+from .models.image_utils import convert_to_rgb
 
 IMAGENET_MEAN = (0.48145466, 0.4578275, 0.40821073)
 IMAGENET_STD = (0.26862954, 0.26130258, 0.27577711)
 
-
-def convert_to_rgb(image):
-    return image.convert("RGB")
-
-
-class LayerScale(nn.Module):
-    def __init__(self, dim, init_values: float = 1e-5, inplace: bool = False):
-        super().__init__()
-        self.weight = nn.Parameter(init_values * torch.ones(dim))
-        self.inplace = inplace
-
-    def forward(self, x):
-        return x.mul_(self.weight) if self.inplace else x * self.weight
+__all__ = [
+    "ImageFeaturesPooler",
+    "VLMConfig",
+    "VLMPreTrainedModel",
+    "VLMForCausalLM",
+    "VLMProcessor",
+]
 
 
 class ImageFeaturesPooler(nn.Module):
