+++ src/uform/models/encoders/__init__.py
@@ -1,0 +1,4 @@
+from .text import TextEncoder, TextEncoderBlock
+from .visual import VisualEncoder, VisualEncoderBlock
+
+__all__ = ["TextEncoder", "TextEncoderBlock", "VisualEncoder", "VisualEncoderBlock"]
