+++ src/uform/models/encoders/text/__init__.py
@@ -1,0 +1,4 @@
+from .block import TextEncoderBlock
+from .encoder import TextEncoder
+
+__all__ = ["TextEncoder", "TextEncoderBlock"]
