+++ src/uform/models/encoders/visual/__init__.py
@@ -1,0 +1,4 @@
+from .block import VisualEncoderBlock
+from .encoder import VisualEncoder
+
+__all__ = ["VisualEncoder", "VisualEncoderBlock"]
