+++ src/uform/models/image_utils.py
@@ -1,0 +1,6 @@
+__all__ = ["convert_to_rgb"]
+
+
+# lambda is not pickable
+def convert_to_rgb(image):
+    return image.convert("RGB")
